pyproject.toml
[tool.poetry]
name = "langgraph-chat-app"
version = "0.1.0"
description = ""
authors = ["Manish Katyan <manishkatyan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
streamlit = "^1.42.2"
langgraph = "0.3.1"
python-dotenv = "^1.0.1"
langchain-openai = "^0.3.7"
typing-extensions = "^4.12.2"
langchain-community = "^0.3.19"
langgraph-prebuilt = "^0.1.1"
tavily-python = "^0.5.1"
chromadb = "^0.6.3"
pypdf = "^5.3.1"
rank-bm25 = "^0.2.2"
langchain-chroma = "^0.2.2"
langchain-cohere = "^0.4.2"
cohere = "^5.14.0"
arize-otel = "^0.8.0"
openinference-instrumentation-langchain = "^0.1.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
